Vector constraints (#148)
* Boundary constraints now correctly support vector constraints.

* _setup_boundary_constraints now raises a DeprecationWarning if the shape of an ODE output is not given in the add_boundary_constraint call.  This will be changed to an exception in Dymos 1.0.0

* Path constraints may now be shaped.
If a path-constrained variable is one of the Dymos known types (state, time, control, parameter, etc) its shape can be inferred automatically.
If a path-constrained variable is an ODE output, we currently cannot infer its shape.  For now, it will be assumed scaler (shape=(1,)) and a DeprecationWarning raised.
Providing an incorrect shape causes issues that will not be obvious for the average user.
In the future, this will be converted to an exception (unless we manage to figure out how to get the shape of ODE outputs prior to setup).