Load case (#236)
* Dymos now requires OpenMDAO version 2.9.0 or later. - Resolves #232
Increased default maxiter for the NewtonSolver when using Pseudospectral transcription with solve_segments=True. - Resolves #234
Added a load_case functionality that will interpolate a previous solution from a CaseReader when the grid has changed. - Resolves #233

The load_case functionality is currently experimental and requires OpenMDAO master as of October 07, 2019 or later.

* pep8 fixes