Implemented control value continuity for ExplicitShooting transcription.
Continuity for ExplicitShooting comes from the default timeseries values.
This means that configure_defects needed to be moved AFTER
configure_timeseries in phase.configure.  Need a lot of checks to make
sure this doesn't break anything.

Reworked ContinuityCompBase to avoid calculation of continuity constraints
when not requested.  This will also need to be propagated through the
other transcriptions.