Changed tooltip hover behavior in bokeh timeseries plots to show data values, no value under cursor.
Addressed timeeseries report test failure that fails if environment variable OPENMDAO_REPORTS=0.
This test now temporarily sets the value to 1 and runs.
Note that this function isn't a true report, although it hijacks the report machinery to place the resulting file in the reports directory.
This function can be run offline from record files, so when reports are not active it assumes that it is running in this mode and outputs the file to the current working directory.