Fixed an issue where default values in add_control or set_control_options were no longer being respected. (#1165)
* Since the control interp comp is no longer in a control group, the set_input_defaults call was no longer meaningful.
These default values had to be applied at add_input instead.

Also, the test to see if default values were sticking was no longer valid because it depended on a problem creation function that was overriding the values after setup anyway.
These issues have been fixed.

* Used set_input_defaults in ExplicitShooting for promoted control values in the phase.
Fixed a numpy deprecation regarding scalar reshapes in reshape_val.
No longer add state and control continuity constraints in ODEEvaluationGroup for the purpose of being able to compute their derivatives when running in simulation and derivatives are not needed.

* fixed tests of ControlInterpComp

* removed unnecessary driver