MultipleShootingUpdateComp was mistakenly applying state rate units to states.
Added time option dt_dstau_targets that allows for dt_dstau to be connected to ODE inputs.
Note that this is experimental behavior and will only work correctly in some transcriptions due to the nature of how the transcriptions work.
Attempting to use ExplicitShooting/simulate with dt_dstau_targets will result in an error, because we currently do not interpolate the value
of dt_dstau smoothly through the phase. In the future, the barycentric control interpolant could do this, but for now it is not possible.