New Radau: Part 2 (#1196)
Added an updated Radau transcription which does not abuse the hybrid implicit/explicit component as we had done previously in the Radau transcription.

This transcription is only available if the environment variable `DYMOS_2='1'` (or some other "truthy" value).
We expect this transcription to fully replace the current Radau transcription but for now it is an experimental feature.
We encourage you to experiment with it and give us feedback.